requirements.txt
hydralit_components==1.0.10
matplotlib==3.7.2
numpy==1.25.2
pandas==2.0.3
plotly==5.16.1
st_clickable_images==0.0.3
streamlit-option-menu==0.3.6
streamlit-js-eval==0.1.5
